+++ src/RcppExports.cpp
@@ -116,27 +116,30 @@ BEGIN_RCPP
 END_RCPP
 }
 // ac_table_colbases
-arma::vec ac_table_colbases(const AcTiterTable titer_table, const std::string min_col_basis, const arma::vec fixed_col_bases);
-RcppExport SEXP _Racmacs_ac_table_colbases(SEXP titer_tableSEXP, SEXP min_col_basisSEXP, SEXP fixed_col_basesSEXP) {
+arma::vec ac_table_colbases(const AcTiterTable titer_table, const std::string min_col_basis, const arma::vec fixed_col_bases, const arma::vec ag_reactivity_adjustments);
+RcppExport SEXP _Racmacs_ac_table_colbases(SEXP titer_tableSEXP, SEXP min_col_basisSEXP, SEXP fixed_col_basesSEXP, SEXP ag_reactivity_adjustmentsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< const AcTiterTable >::type titer_table(titer_tableSEXP);
     Rcpp::traits::input_parameter< const std::string >::type min_col_basis(min_col_basisSEXP);
     Rcpp::traits::input_parameter< const arma::vec >::type fixed_col_bases(fixed_col_basesSEXP);
-    rcpp_result_gen = Rcpp::wrap(ac_table_colbases(titer_table, min_col_basis, fixed_col_bases));
+    Rcpp::traits::input_parameter< const arma::vec >::type ag_reactivity_adjustments(ag_reactivity_adjustmentsSEXP);
+    rcpp_result_gen = Rcpp::wrap(ac_table_colbases(titer_table, min_col_basis, fixed_col_bases, ag_reactivity_adjustments));
     return rcpp_result_gen;
 END_RCPP
 }
 // ac_numeric_table_distances
-arma::mat ac_numeric_table_distances(const AcTiterTable titer_table, const arma::vec colbases);
-RcppExport SEXP _Racmacs_ac_numeric_table_distances(SEXP titer_tableSEXP, SEXP colbasesSEXP) {
+arma::mat ac_numeric_table_distances(const AcTiterTable titer_table, const std::string min_col_basis, const arma::vec fixed_col_bases, const arma::vec ag_reactivity_adjustments);
+RcppExport SEXP _Racmacs_ac_numeric_table_distances(SEXP titer_tableSEXP, SEXP min_col_basisSEXP, SEXP fixed_col_basesSEXP, SEXP ag_reactivity_adjustmentsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< const AcTiterTable >::type titer_table(titer_tableSEXP);
-    Rcpp::traits::input_parameter< const arma::vec >::type colbases(colbasesSEXP);
-    rcpp_result_gen = Rcpp::wrap(ac_numeric_table_distances(titer_table, colbases));
+    Rcpp::traits::input_parameter< const std::string >::type min_col_basis(min_col_basisSEXP);
+    Rcpp::traits::input_parameter< const arma::vec >::type fixed_col_bases(fixed_col_basesSEXP);
+    Rcpp::traits::input_parameter< const arma::vec >::type ag_reactivity_adjustments(ag_reactivity_adjustmentsSEXP);
+    rcpp_result_gen = Rcpp::wrap(ac_numeric_table_distances(titer_table, min_col_basis, fixed_col_bases, ag_reactivity_adjustments));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -170,8 +173,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // ac_optimize_map
-AcMap ac_optimize_map(AcMap map, int num_dims, int num_optimizations, std::string min_col_basis, arma::vec fixed_col_bases, arma::mat titer_weights, AcOptimizerOptions options);
-RcppExport SEXP _Racmacs_ac_optimize_map(SEXP mapSEXP, SEXP num_dimsSEXP, SEXP num_optimizationsSEXP, SEXP min_col_basisSEXP, SEXP fixed_col_basesSEXP, SEXP titer_weightsSEXP, SEXP optionsSEXP) {
+AcMap ac_optimize_map(AcMap map, int num_dims, int num_optimizations, std::string min_col_basis, arma::vec fixed_col_bases, arma::vec ag_reactivity_adjustments, arma::mat titer_weights, AcOptimizerOptions options);
+RcppExport SEXP _Racmacs_ac_optimize_map(SEXP mapSEXP, SEXP num_dimsSEXP, SEXP num_optimizationsSEXP, SEXP min_col_basisSEXP, SEXP fixed_col_basesSEXP, SEXP ag_reactivity_adjustmentsSEXP, SEXP titer_weightsSEXP, SEXP optionsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -180,9 +183,10 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< int >::type num_optimizations(num_optimizationsSEXP);
     Rcpp::traits::input_parameter< std::string >::type min_col_basis(min_col_basisSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type fixed_col_bases(fixed_col_basesSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type ag_reactivity_adjustments(ag_reactivity_adjustmentsSEXP);
     Rcpp::traits::input_parameter< arma::mat >::type titer_weights(titer_weightsSEXP);
     Rcpp::traits::input_parameter< AcOptimizerOptions >::type options(optionsSEXP);
-    rcpp_result_gen = Rcpp::wrap(ac_optimize_map(map, num_dims, num_optimizations, min_col_basis, fixed_col_bases, titer_weights, options));
+    rcpp_result_gen = Rcpp::wrap(ac_optimize_map(map, num_dims, num_optimizations, min_col_basis, fixed_col_bases, ag_reactivity_adjustments, titer_weights, options));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -466,6 +470,16 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// ac_opt_get_agreactivityadjustments
+arma::vec ac_opt_get_agreactivityadjustments(const AcOptimization opt);
+RcppExport SEXP _Racmacs_ac_opt_get_agreactivityadjustments(SEXP optSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< const AcOptimization >::type opt(optSEXP);
+    rcpp_result_gen = Rcpp::wrap(ac_opt_get_agreactivityadjustments(opt));
+    return rcpp_result_gen;
+END_RCPP
+}
 // ac_opt_get_stress
 double ac_opt_get_stress(const AcOptimization opt);
 RcppExport SEXP _Racmacs_ac_opt_get_stress(SEXP optSEXP) {
@@ -559,6 +573,17 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< AcOptimization >::type opt(optSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type value(valueSEXP);
     rcpp_result_gen = Rcpp::wrap(ac_opt_set_fixedcolbases(opt, value));
+    return rcpp_result_gen;
+END_RCPP
+}
+// ac_opt_set_agreactivityadjustments
+AcOptimization ac_opt_set_agreactivityadjustments(AcOptimization opt, arma::vec value);
+RcppExport SEXP _Racmacs_ac_opt_set_agreactivityadjustments(SEXP optSEXP, SEXP valueSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< AcOptimization >::type opt(optSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type value(valueSEXP);
+    rcpp_result_gen = Rcpp::wrap(ac_opt_set_agreactivityadjustments(opt, value));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -1018,8 +1043,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // ac_bootstrap_map
-BootstrapOutput ac_bootstrap_map(AcMap map, std::string method, bool bootstrap_ags, bool bootstrap_sr, bool reoptimize, double ag_noise_sd, double titer_noise_sd, std::string minimum_column_basis, arma::vec fixed_column_bases, int num_optimizations, int num_dimensions, AcOptimizerOptions options);
-RcppExport SEXP _Racmacs_ac_bootstrap_map(SEXP mapSEXP, SEXP methodSEXP, SEXP bootstrap_agsSEXP, SEXP bootstrap_srSEXP, SEXP reoptimizeSEXP, SEXP ag_noise_sdSEXP, SEXP titer_noise_sdSEXP, SEXP minimum_column_basisSEXP, SEXP fixed_column_basesSEXP, SEXP num_optimizationsSEXP, SEXP num_dimensionsSEXP, SEXP optionsSEXP) {
+BootstrapOutput ac_bootstrap_map(AcMap map, std::string method, bool bootstrap_ags, bool bootstrap_sr, bool reoptimize, double ag_noise_sd, double titer_noise_sd, std::string minimum_column_basis, arma::vec fixed_column_bases, arma::vec ag_reactivity_adjustments, int num_optimizations, int num_dimensions, AcOptimizerOptions options);
+RcppExport SEXP _Racmacs_ac_bootstrap_map(SEXP mapSEXP, SEXP methodSEXP, SEXP bootstrap_agsSEXP, SEXP bootstrap_srSEXP, SEXP reoptimizeSEXP, SEXP ag_noise_sdSEXP, SEXP titer_noise_sdSEXP, SEXP minimum_column_basisSEXP, SEXP fixed_column_basesSEXP, SEXP ag_reactivity_adjustmentsSEXP, SEXP num_optimizationsSEXP, SEXP num_dimensionsSEXP, SEXP optionsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -1032,16 +1057,17 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< double >::type titer_noise_sd(titer_noise_sdSEXP);
     Rcpp::traits::input_parameter< std::string >::type minimum_column_basis(minimum_column_basisSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type fixed_column_bases(fixed_column_basesSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type ag_reactivity_adjustments(ag_reactivity_adjustmentsSEXP);
     Rcpp::traits::input_parameter< int >::type num_optimizations(num_optimizationsSEXP);
     Rcpp::traits::input_parameter< int >::type num_dimensions(num_dimensionsSEXP);
     Rcpp::traits::input_parameter< AcOptimizerOptions >::type options(optionsSEXP);
-    rcpp_result_gen = Rcpp::wrap(ac_bootstrap_map(map, method, bootstrap_ags, bootstrap_sr, reoptimize, ag_noise_sd, titer_noise_sd, minimum_column_basis, fixed_column_bases, num_optimizations, num_dimensions, options));
+    rcpp_result_gen = Rcpp::wrap(ac_bootstrap_map(map, method, bootstrap_ags, bootstrap_sr, reoptimize, ag_noise_sd, titer_noise_sd, minimum_column_basis, fixed_column_bases, ag_reactivity_adjustments, num_optimizations, num_dimensions, options));
     return rcpp_result_gen;
 END_RCPP
 }
 // ac_dimension_test_map
-DimTestOutput ac_dimension_test_map(AcTiterTable titer_table, arma::uvec dimensions_to_test, double test_proportion, std::string minimum_column_basis, arma::vec fixed_column_bases, int num_optimizations, AcOptimizerOptions options);
-RcppExport SEXP _Racmacs_ac_dimension_test_map(SEXP titer_tableSEXP, SEXP dimensions_to_testSEXP, SEXP test_proportionSEXP, SEXP minimum_column_basisSEXP, SEXP fixed_column_basesSEXP, SEXP num_optimizationsSEXP, SEXP optionsSEXP) {
+DimTestOutput ac_dimension_test_map(AcTiterTable titer_table, arma::uvec dimensions_to_test, double test_proportion, std::string minimum_column_basis, arma::vec fixed_column_bases, arma::vec ag_reactivity_adjustments, int num_optimizations, AcOptimizerOptions options);
+RcppExport SEXP _Racmacs_ac_dimension_test_map(SEXP titer_tableSEXP, SEXP dimensions_to_testSEXP, SEXP test_proportionSEXP, SEXP minimum_column_basisSEXP, SEXP fixed_column_basesSEXP, SEXP ag_reactivity_adjustmentsSEXP, SEXP num_optimizationsSEXP, SEXP optionsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -1050,25 +1076,25 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< double >::type test_proportion(test_proportionSEXP);
     Rcpp::traits::input_parameter< std::string >::type minimum_column_basis(minimum_column_basisSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type fixed_column_bases(fixed_column_basesSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type ag_reactivity_adjustments(ag_reactivity_adjustmentsSEXP);
     Rcpp::traits::input_parameter< int >::type num_optimizations(num_optimizationsSEXP);
     Rcpp::traits::input_parameter< AcOptimizerOptions >::type options(optionsSEXP);
-    rcpp_result_gen = Rcpp::wrap(ac_dimension_test_map(titer_table, dimensions_to_test, test_proportion, minimum_column_basis, fixed_column_bases, num_optimizations, options));
+    rcpp_result_gen = Rcpp::wrap(ac_dimension_test_map(titer_table, dimensions_to_test, test_proportion, minimum_column_basis, fixed_column_bases, ag_reactivity_adjustments, num_optimizations, options));
     return rcpp_result_gen;
 END_RCPP
 }
 // ac_hemi_test
-AcOptimization ac_hemi_test(AcOptimization optimization, AcTiterTable titertable, arma::vec colbases, double grid_spacing, double stress_lim, AcOptimizerOptions options);
-RcppExport SEXP _Racmacs_ac_hemi_test(SEXP optimizationSEXP, SEXP titertableSEXP, SEXP colbasesSEXP, SEXP grid_spacingSEXP, SEXP stress_limSEXP, SEXP optionsSEXP) {
+AcOptimization ac_hemi_test(AcOptimization optimization, AcTiterTable titertable, double grid_spacing, double stress_lim, AcOptimizerOptions options);
+RcppExport SEXP _Racmacs_ac_hemi_test(SEXP optimizationSEXP, SEXP titertableSEXP, SEXP grid_spacingSEXP, SEXP stress_limSEXP, SEXP optionsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< AcOptimization >::type optimization(optimizationSEXP);
     Rcpp::traits::input_parameter< AcTiterTable >::type titertable(titertableSEXP);
-    Rcpp::traits::input_parameter< arma::vec >::type colbases(colbasesSEXP);
     Rcpp::traits::input_parameter< double >::type grid_spacing(grid_spacingSEXP);
     Rcpp::traits::input_parameter< double >::type stress_lim(stress_limSEXP);
     Rcpp::traits::input_parameter< AcOptimizerOptions >::type options(optionsSEXP);
-    rcpp_result_gen = Rcpp::wrap(ac_hemi_test(optimization, titertable, colbases, grid_spacing, stress_lim, options));
+    rcpp_result_gen = Rcpp::wrap(ac_hemi_test(optimization, titertable, grid_spacing, stress_lim, options));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -1195,32 +1221,33 @@ BEGIN_RCPP
 END_RCPP
 }
 // ac_move_trapped_points
-AcOptimization ac_move_trapped_points(AcOptimization optimization, AcTiterTable titertable, arma::vec colbases, double grid_spacing, AcOptimizerOptions options, int max_iterations);
-RcppExport SEXP _Racmacs_ac_move_trapped_points(SEXP optimizationSEXP, SEXP titertableSEXP, SEXP colbasesSEXP, SEXP grid_spacingSEXP, SEXP optionsSEXP, SEXP max_iterationsSEXP) {
+AcOptimization ac_move_trapped_points(AcOptimization optimization, AcTiterTable titertable, double grid_spacing, AcOptimizerOptions options, int max_iterations);
+RcppExport SEXP _Racmacs_ac_move_trapped_points(SEXP optimizationSEXP, SEXP titertableSEXP, SEXP grid_spacingSEXP, SEXP optionsSEXP, SEXP max_iterationsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< AcOptimization >::type optimization(optimizationSEXP);
     Rcpp::traits::input_parameter< AcTiterTable >::type titertable(titertableSEXP);
-    Rcpp::traits::input_parameter< arma::vec >::type colbases(colbasesSEXP);
     Rcpp::traits::input_parameter< double >::type grid_spacing(grid_spacingSEXP);
     Rcpp::traits::input_parameter< AcOptimizerOptions >::type options(optionsSEXP);
     Rcpp::traits::input_parameter< int >::type max_iterations(max_iterationsSEXP);
-    rcpp_result_gen = Rcpp::wrap(ac_move_trapped_points(optimization, titertable, colbases, grid_spacing, options, max_iterations));
+    rcpp_result_gen = Rcpp::wrap(ac_move_trapped_points(optimization, titertable, grid_spacing, options, max_iterations));
     return rcpp_result_gen;
 END_RCPP
 }
 // ac_coords_stress
-double ac_coords_stress(const AcTiterTable& titers, const arma::vec& colbases, arma::mat& ag_coords, arma::mat& sr_coords);
-RcppExport SEXP _Racmacs_ac_coords_stress(SEXP titersSEXP, SEXP colbasesSEXP, SEXP ag_coordsSEXP, SEXP sr_coordsSEXP) {
+double ac_coords_stress(const AcTiterTable& titers, const std::string& min_colbasis, const arma::vec& fixed_colbases, const arma::vec& ag_reactivity_adjustments, arma::mat& ag_coords, arma::mat& sr_coords);
+RcppExport SEXP _Racmacs_ac_coords_stress(SEXP titersSEXP, SEXP min_colbasisSEXP, SEXP fixed_colbasesSEXP, SEXP ag_reactivity_adjustmentsSEXP, SEXP ag_coordsSEXP, SEXP sr_coordsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< const AcTiterTable& >::type titers(titersSEXP);
-    Rcpp::traits::input_parameter< const arma::vec& >::type colbases(colbasesSEXP);
+    Rcpp::traits::input_parameter< const std::string& >::type min_colbasis(min_colbasisSEXP);
+    Rcpp::traits::input_parameter< const arma::vec& >::type fixed_colbases(fixed_colbasesSEXP);
+    Rcpp::traits::input_parameter< const arma::vec& >::type ag_reactivity_adjustments(ag_reactivity_adjustmentsSEXP);
     Rcpp::traits::input_parameter< arma::mat& >::type ag_coords(ag_coordsSEXP);
     Rcpp::traits::input_parameter< arma::mat& >::type sr_coords(sr_coordsSEXP);
-    rcpp_result_gen = Rcpp::wrap(ac_coords_stress(titers, colbases, ag_coords, sr_coords));
+    rcpp_result_gen = Rcpp::wrap(ac_coords_stress(titers, min_colbasis, fixed_colbases, ag_reactivity_adjustments, ag_coords, sr_coords));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -1243,18 +1270,20 @@ BEGIN_RCPP
 END_RCPP
 }
 // ac_runOptimizations
-std::vector<AcOptimization> ac_runOptimizations(const AcTiterTable& titertable, const arma::vec& colbases, const arma::uword& num_dims, const arma::uword& num_optimizations, const AcOptimizerOptions& options, const arma::mat& titer_weights);
-RcppExport SEXP _Racmacs_ac_runOptimizations(SEXP titertableSEXP, SEXP colbasesSEXP, SEXP num_dimsSEXP, SEXP num_optimizationsSEXP, SEXP optionsSEXP, SEXP titer_weightsSEXP) {
+std::vector<AcOptimization> ac_runOptimizations(const AcTiterTable& titertable, const std::string& minimum_col_basis, const arma::vec& fixed_colbases, const arma::vec& ag_reactivity_adjustments, const arma::uword& num_dims, const arma::uword& num_optimizations, const AcOptimizerOptions& options, const arma::mat& titer_weights);
+RcppExport SEXP _Racmacs_ac_runOptimizations(SEXP titertableSEXP, SEXP minimum_col_basisSEXP, SEXP fixed_colbasesSEXP, SEXP ag_reactivity_adjustmentsSEXP, SEXP num_dimsSEXP, SEXP num_optimizationsSEXP, SEXP optionsSEXP, SEXP titer_weightsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< const AcTiterTable& >::type titertable(titertableSEXP);
-    Rcpp::traits::input_parameter< const arma::vec& >::type colbases(colbasesSEXP);
+    Rcpp::traits::input_parameter< const std::string& >::type minimum_col_basis(minimum_col_basisSEXP);
+    Rcpp::traits::input_parameter< const arma::vec& >::type fixed_colbases(fixed_colbasesSEXP);
+    Rcpp::traits::input_parameter< const arma::vec& >::type ag_reactivity_adjustments(ag_reactivity_adjustmentsSEXP);
     Rcpp::traits::input_parameter< const arma::uword& >::type num_dims(num_dimsSEXP);
     Rcpp::traits::input_parameter< const arma::uword& >::type num_optimizations(num_optimizationsSEXP);
     Rcpp::traits::input_parameter< const AcOptimizerOptions& >::type options(optionsSEXP);
     Rcpp::traits::input_parameter< const arma::mat& >::type titer_weights(titer_weightsSEXP);
-    rcpp_result_gen = Rcpp::wrap(ac_runOptimizations(titertable, colbases, num_dims, num_optimizations, options, titer_weights));
+    rcpp_result_gen = Rcpp::wrap(ac_runOptimizations(titertable, minimum_col_basis, fixed_colbases, ag_reactivity_adjustments, num_dims, num_optimizations, options, titer_weights));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -1443,11 +1472,11 @@ static const R_CallMethodDef CallEntries[] = {
     {"_Racmacs_ac_align_map", (DL_FUNC) &_Racmacs_ac_align_map, 4},
     {"_Racmacs_ac_align_optimizations", (DL_FUNC) &_Racmacs_ac_align_optimizations, 1},
     {"_Racmacs_ac_subset_map", (DL_FUNC) &_Racmacs_ac_subset_map, 3},
-    {"_Racmacs_ac_table_colbases", (DL_FUNC) &_Racmacs_ac_table_colbases, 3},
-    {"_Racmacs_ac_numeric_table_distances", (DL_FUNC) &_Racmacs_ac_numeric_table_distances, 2},
+    {"_Racmacs_ac_table_colbases", (DL_FUNC) &_Racmacs_ac_table_colbases, 4},
+    {"_Racmacs_ac_numeric_table_distances", (DL_FUNC) &_Racmacs_ac_numeric_table_distances, 4},
     {"_Racmacs_ac_newOptimization", (DL_FUNC) &_Racmacs_ac_newOptimization, 3},
     {"_Racmacs_ac_relaxOptimization", (DL_FUNC) &_Racmacs_ac_relaxOptimization, 6},
-    {"_Racmacs_ac_optimize_map", (DL_FUNC) &_Racmacs_ac_optimize_map, 7},
+    {"_Racmacs_ac_optimize_map", (DL_FUNC) &_Racmacs_ac_optimize_map, 8},
     {"_Racmacs_ac_new_antigen", (DL_FUNC) &_Racmacs_ac_new_antigen, 1},
     {"_Racmacs_ac_ag_get_id", (DL_FUNC) &_Racmacs_ac_ag_get_id, 1},
     {"_Racmacs_ac_ag_get_date", (DL_FUNC) &_Racmacs_ac_ag_get_date, 1},
@@ -1475,6 +1504,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_Racmacs_ac_opt_get_translation", (DL_FUNC) &_Racmacs_ac_opt_get_translation, 1},
     {"_Racmacs_ac_opt_get_mincolbasis", (DL_FUNC) &_Racmacs_ac_opt_get_mincolbasis, 1},
     {"_Racmacs_ac_opt_get_fixedcolbases", (DL_FUNC) &_Racmacs_ac_opt_get_fixedcolbases, 1},
+    {"_Racmacs_ac_opt_get_agreactivityadjustments", (DL_FUNC) &_Racmacs_ac_opt_get_agreactivityadjustments, 1},
     {"_Racmacs_ac_opt_get_stress", (DL_FUNC) &_Racmacs_ac_opt_get_stress, 1},
     {"_Racmacs_ac_opt_get_dimensions", (DL_FUNC) &_Racmacs_ac_opt_get_dimensions, 1},
     {"_Racmacs_ac_opt_get_comment", (DL_FUNC) &_Racmacs_ac_opt_get_comment, 1},
@@ -1484,6 +1514,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_Racmacs_ac_opt_set_translation", (DL_FUNC) &_Racmacs_ac_opt_set_translation, 2},
     {"_Racmacs_ac_opt_set_mincolbasis", (DL_FUNC) &_Racmacs_ac_opt_set_mincolbasis, 2},
     {"_Racmacs_ac_opt_set_fixedcolbases", (DL_FUNC) &_Racmacs_ac_opt_set_fixedcolbases, 2},
+    {"_Racmacs_ac_opt_set_agreactivityadjustments", (DL_FUNC) &_Racmacs_ac_opt_set_agreactivityadjustments, 2},
     {"_Racmacs_ac_opt_set_stress", (DL_FUNC) &_Racmacs_ac_opt_set_stress, 2},
     {"_Racmacs_ac_opt_set_comment", (DL_FUNC) &_Racmacs_ac_opt_set_comment, 2},
     {"_Racmacs_ac_rotate_optimization", (DL_FUNC) &_Racmacs_ac_rotate_optimization, 3},
@@ -1527,9 +1558,9 @@ static const R_CallMethodDef CallEntries[] = {
     {"_Racmacs_ac_sr_set_name_abbreviated", (DL_FUNC) &_Racmacs_ac_sr_set_name_abbreviated, 2},
     {"_Racmacs_ac_sr_set_group", (DL_FUNC) &_Racmacs_ac_sr_set_group, 2},
     {"_Racmacs_ac_sr_set_group_levels", (DL_FUNC) &_Racmacs_ac_sr_set_group_levels, 2},
-    {"_Racmacs_ac_bootstrap_map", (DL_FUNC) &_Racmacs_ac_bootstrap_map, 12},
-    {"_Racmacs_ac_dimension_test_map", (DL_FUNC) &_Racmacs_ac_dimension_test_map, 7},
-    {"_Racmacs_ac_hemi_test", (DL_FUNC) &_Racmacs_ac_hemi_test, 6},
+    {"_Racmacs_ac_bootstrap_map", (DL_FUNC) &_Racmacs_ac_bootstrap_map, 13},
+    {"_Racmacs_ac_dimension_test_map", (DL_FUNC) &_Racmacs_ac_dimension_test_map, 8},
+    {"_Racmacs_ac_hemi_test", (DL_FUNC) &_Racmacs_ac_hemi_test, 5},
     {"_Racmacs_ac_match_map_ags", (DL_FUNC) &_Racmacs_ac_match_map_ags, 2},
     {"_Racmacs_ac_match_map_sr", (DL_FUNC) &_Racmacs_ac_match_map_sr, 2},
     {"_Racmacs_ac_merge_titer_layers", (DL_FUNC) &_Racmacs_ac_merge_titer_layers, 1},
@@ -1540,10 +1571,10 @@ static const R_CallMethodDef CallEntries[] = {
     {"_Racmacs_ac_merge_frozen_merge", (DL_FUNC) &_Racmacs_ac_merge_frozen_merge, 2},
     {"_Racmacs_ac_merge_incremental", (DL_FUNC) &_Racmacs_ac_merge_incremental, 5},
     {"_Racmacs_ac_merge_titers", (DL_FUNC) &_Racmacs_ac_merge_titers, 2},
-    {"_Racmacs_ac_move_trapped_points", (DL_FUNC) &_Racmacs_ac_move_trapped_points, 6},
-    {"_Racmacs_ac_coords_stress", (DL_FUNC) &_Racmacs_ac_coords_stress, 4},
+    {"_Racmacs_ac_move_trapped_points", (DL_FUNC) &_Racmacs_ac_move_trapped_points, 5},
+    {"_Racmacs_ac_coords_stress", (DL_FUNC) &_Racmacs_ac_coords_stress, 6},
     {"_Racmacs_ac_relax_coords", (DL_FUNC) &_Racmacs_ac_relax_coords, 8},
-    {"_Racmacs_ac_runOptimizations", (DL_FUNC) &_Racmacs_ac_runOptimizations, 6},
+    {"_Racmacs_ac_runOptimizations", (DL_FUNC) &_Racmacs_ac_runOptimizations, 8},
     {"_Racmacs_ac_stress_blob_grid", (DL_FUNC) &_Racmacs_ac_stress_blob_grid, 6},
     {"_Racmacs_numeric_titers", (DL_FUNC) &_Racmacs_numeric_titers, 1},
     {"_Racmacs_log_titers", (DL_FUNC) &_Racmacs_log_titers, 1},