+++ src/ac_optim_map_stress.h
@@ -9,9 +9,11 @@
 
 // Generating optimizations with randomised coords
 std::vector<AcOptimization> ac_generateOptimizations(
-    const arma::vec &colbases,
     const arma::mat &tabledist_matrix,
     const arma::umat &titertype_matrix,
+    const std::string &min_colbasis,
+    const arma::vec &fixed_colbases,
+    const arma::vec &ag_reactivity_adjustments,
     const int &num_dims,
     const int &num_optimizations,
     const AcOptimizerOptions &options
@@ -20,7 +22,6 @@ std::vector<AcOptimization> ac_generateOptimizations(
 // Relaxing optimizations
 void ac_relaxOptimizations(
     std::vector<AcOptimization>& optimizations,
-    const arma::vec &colbases,
     const arma::mat &tabledist_matrix,
     const arma::umat &titertype_matrix,
     const AcOptimizerOptions &options,
@@ -30,7 +31,9 @@ void ac_relaxOptimizations(
 // Running optimizations
 std::vector<AcOptimization> ac_runOptimizations(
     const AcTiterTable &titertable,
-    const arma::vec &colbases,
+    const std::string &min_colbasis,
+    const arma::vec &fixed_colbases,
+    const arma::vec &ag_reactivity_adjustments,
     const arma::uword &num_dims,
     const arma::uword &num_optimizations,
     const AcOptimizerOptions &options,