+++ NEWS.md
@@ -1,4 +1,5 @@
-# Racmacs 1.1.10
+# Racmacs 1.1.11
 
 * Added a `NEWS.md` file to track changes to the package.
-* Add test and correction for error where coordinates of under-constrained points were set to NaN for the first optimization run only, not all.
+* Added test and correction for error where coordinates of under-constrained points were set to NaN for the first optimization run only, not all.
+* Added integrated support for setting antigen reactivity adjustments